+++ PayPalWebCheckout/src/main/java/com/paypal/android/checkoutweb/PayPalWebCheckoutResult.kt
@@ -1,6 +1,0 @@
-package com.paypal.android.checkoutweb
-
-/**
- * A result passed to a [PayPalWebCheckoutListener] when the PayPal flow completes successfully.
- */
-data class PayPalWebCheckoutResult(val orderId: String?, val payerId: String?)