+++ PayPalWebPayments/src/main/java/com/paypal/android/paypalwebpayments/PayPalWebCheckoutRequest.kt
@@ -1,0 +1,12 @@
+package com.paypal.android.paypalwebpayments
+
+/**
+ * Creates an instance of a PayPalRequest.
+ *
+ * @param orderID The ID of the order to be approved.
+ * @param fundingSource specify funding (credit, paylater or default)
+ */
+data class PayPalWebCheckoutRequest @JvmOverloads constructor (
+    val orderID: String,
+    val fundingSource: PayPalWebCheckoutFundingSource = PayPalWebCheckoutFundingSource.PAYPAL
+)